third_party/WebKit/LayoutTests/fast/css/invalidation/font-feature-settings.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<style>
    #x {
        -webkit-font-feature-settings: "liga";
        font-feature-settings: "liga"
    }
    .y {
        background-color: green
    }
</style>
<div id="x">
    <div>
        <div></div>
        <div></div>
        <div></div>
        <div></div>
    </div>
</div>
<script>
    description("Local style change on element with font-feature-settings should not cause subtree recalc");

    x.offsetTop;
    x.className = "y";

    if (window.internals)
        shouldBe("internals.updateStyleAndReturnAffectedElementCount()", "1");

</script>
