third_party/blink/web_tests/platform/generic/fast/css/invalidation/font-feature-settings-expected.txt

Local style change on element with font-feature-settings should not cause subtree recalc

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS successfullyParsed is true

TEST COMPLETE
